Adding sounds to config and input tracer form
-While configuring bells, a sound is played when the bell rings.
-Input tracer form is now available from the main menu
-Input tracer latches onto whatever pin is triggered first once it has
opened and displays the input from that pin on the screen
-Adding a debounce value will put a debounce setting in. This should be
adjusted so that a single square wave appears for each trigger and
release of the input
-Clicking the hold options will stop the graph moving with time and
instead align each input with the right hand side. This mode also gives
an estimate of the width of the signal, to give an idea of what the
debounce value should be.